Refactorings + split vertex stuff into separate namespace/files.
