+++ osu.Framework/Graphics/Containers/BufferedContainerDrawNode.cs
@@ -15,6 +15,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics.Shaders;
 using System;
 using osu.Framework.Graphics.Colour;
+using osu.Framework.Graphics.OpenGL.Vertices;
 
 namespace osu.Framework.Graphics.Containers
 {
