+++ osu.Framework/Graphics/DrawNode.cs
@@ -3,6 +3,7 @@
 
 using osu.Framework.Graphics.OpenGL;
 using System;
+using osu.Framework.Graphics.OpenGL.Vertices;
 
 namespace osu.Framework.Graphics
 {
