+++ osu.Framework/Graphics/Lines/PathDrawNode.cs
@@ -10,6 +10,7 @@ using OpenTK;
 using System;
 using System.Collections.Generic;
 using osu.Framework.Graphics.Batches;
+using osu.Framework.Graphics.OpenGL.Vertices;
 using OpenTK.Graphics;
 
 namespace osu.Framework.Graphics.Lines
