+++ osu.Framework/Graphics/OpenGL/Buffers/VertexBuffer.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using System;
-using System.Reflection;
+using osu.Framework.Graphics.OpenGL.Vertices;
 using OpenTK.Graphics.ES30;
 using osu.Framework.Statistics;
