+++ osu.Framework/Graphics/OpenGL/Textures/TextureGL.cs
@@ -6,6 +6,7 @@ using osu.Framework.Graphics.Primitives;
 using OpenTK.Graphics.ES30;
 using OpenTK;
 using osu.Framework.Graphics.Colour;
+using osu.Framework.Graphics.OpenGL.Vertices;
 
 namespace osu.Framework.Graphics.OpenGL.Textures
 {
