+++ osu.Framework/Graphics/OpenGL/Vertices/IVertex.cs
@@ -1,0 +1,9 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+namespace osu.Framework.Graphics.OpenGL.Vertices
+{
+    public interface IVertex
+    {
+    }
+}
