+++ osu.Framework/Graphics/OpenGL/Vertices/Vertex2D.cs
@@ -1,0 +1,25 @@
+﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+using System;
+using System.Runtime.InteropServices;
+using OpenTK;
+using OpenTK.Graphics;
+using OpenTK.Graphics.ES30;
+
+namespace osu.Framework.Graphics.OpenGL.Vertices
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct Vertex2D : IEquatable<Vertex2D>, IVertex
+    {
+        [VertexMember(2, VertexAttribPointerType.Float)]
+        public Vector2 Position;
+        [VertexMember(4, VertexAttribPointerType.Float)]
+        public Color4 Colour;
+
+        public bool Equals(Vertex2D other)
+        {
+            return Position.Equals(other.Position) && Colour.Equals(other.Colour);
+        }
+    }
+}
