+++ osu.Framework/Graphics/OpenGL/Vertices/VertexMemberAttribute.cs
@@ -1,0 +1,49 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+using System;
+using OpenTK.Graphics.ES30;
+
+namespace osu.Framework.Graphics.OpenGL.Vertices
+{
+    [AttributeUsage(AttributeTargets.Field)]
+    public class VertexMemberAttribute : Attribute
+    {
+        /// <summary>
+        /// The number of components of <see cref="Type"/> represented by this attribute.
+        /// E.g. an <see cref="OpenTK.Vector2"/> is represented by **2** <see cref="VertexAttribPointerType.Float"/> components.
+        /// </summary>
+        public int Count { get; private set; }
+
+        /// <summary>
+        /// The type of this vertex member. This should reflect the type of data represented in the shader.
+        /// E.g. an <see cref="OpenTK.Vector2"/> is represented by 2 **<see cref="VertexAttribPointerType.Float"/>** components.
+        /// </summary>
+        public VertexAttribPointerType Type { get; private set; }
+
+        /// <summary>
+        /// Whether this vertex member value is normalized. If this is set to true, the data will be mapped to a range of [-1, 1] (signed) or [0, 1] (unsigned)
+        /// when it is passed to the shader.
+        /// </summary>
+        public bool Normalized { get; private set; }
+
+        /// <summary>
+        /// The offset of this attribute member in the struct. This is computed internally by the framework.
+        /// </summary>
+        internal IntPtr Offset;
+
+        public VertexMemberAttribute(int count, VertexAttribPointerType type)
+        {
+            Count = count;
+            Type = type;
+            Normalized = false;
+        }
+
+        public VertexMemberAttribute(int count, VertexAttribPointerType type, bool normalized)
+        {
+            Count = count;
+            Type = type;
+            Normalized = normalized;
+        }
+    }
+}
